Alzheimer-sDiseaseClassifier-code/requirements.txt
tensorflow==2.9.1
streamlit-lottie==0.0.3
streamlit==1.9.0
requests==2.26.0
keras==2.9.0
Pillow==9.2.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
opencv-python==4.6.0.66
numpy==1.21.2



py -m streamlit run app.py